requirements.txt
django==4.1
django-ckeditor==6.5.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
